+++ lib/presentation/splash_page/ui/view/splash_page.dart
